# removing show loading status of patterns from file