by showing loading status of patterns from file taking extra 5 to 15 seconds time to start the program based on number of patterns (ie: 100000 patterns).
so it's better to remove that the showing patterns to make program to start faster while using patterns from file to 1 to 3 seconds only.

please accept it.
Thank you.