@exploitagency 2018-08-27
Can we add a flag to enable or disable this?
Some people specifically wanted this option.